# docs/swagger.yaml
openapi: 3.0.0
servers:
  - url: localhost:25565
info:
  description: An API used to do banking operations such as user register, withdrawl, deposit, access control, internal and external accounts transfers, account blocking and more.
  version: "1.0.0"
  title: "Simple Banking API"
  contact:
    email: julio.kobayashi@ebury.com
  license:
    name: Apache 2.0
    url: "[http://www.apache.org/licenses/LICENSE-2.0.html](http://www.apache.org/licenses/LICENSE-2.0.html)"
...
paths:
...
  /clients:
    post:
      tags:
        - clients
      summary: creates a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
      responses:
        "200":
          description: client created sucessfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        "400":
          description: bad input parameter
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
  /clients/{client_id}:
    get:
      tags:
        - clients
      summary: return data of client with ID clientID
      parameters:
        - in: path
          name: client_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: client returned sucessfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
    delete:
      tags:
        - clients
      summary: delete client of clientID
      parameters:
        - in: path
          name: client_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client deleted seccesfully
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
    put:
      tags:
        - clients
      summary: update a given piece of information of a certain client
      parameters:
        - in: path
          name: client_id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
      responses:
        "200":
          description: Client updated sucessfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
  /clients/{client_id}/report:
    get:
      tags:
        - clients
      summary: returns a report of the client
      parameters:
        - in: path
          name: client_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientReport'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
  /clients/{client_id}/block:
    put:
      tags:
        - clients
      summary: block a client
      parameters:
        - in: path
          name: client_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client Blocked
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
  /clients/{client_id}/unblock:
    put:
      tags:
        - clients
      summary: unblock a client
      parameters:
        - in: path
          name: client_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Client Unblocked
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
...
  /transfers:
    post:
      tags:
        - transfers
      summary: creates a transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Transfer'
      responses:
        "200":
          description: Transfer created sucessfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
  /transfers/{transfer_id}:
    get:
      tags:
        - transfers
      summary: return data of transfer with ID transferID
      parameters:
        - in: path
          name: transfer_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: transfer returned sucessfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transfer'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
    delete:
      tags:
        - transfers
      summary: delete transfer of transferID
      parameters:
        - in: path
          name: transfer_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transfer deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
  /deposits:
    post:
      tags:
        - deposits
      summary: creates a deposit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Deposit'
      responses:
        "201":
          description: Deposit created sucessfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deposit'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
  /deposits/{deposit_id}:
    get:
      tags:
        - deposits
      summary: return data of deposit with ID depositID
      parameters:
        - in: path
          name: deposit_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: deposit returned sucessfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deposit'
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
    delete:
      tags:
        - deposits
      summary: delete deposit of depositID
      parameters:
        - in: path
          name: deposit_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deposit deleted succesfully!
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "500":
          description: Internal Server Error
...
components:
  schemas:
...
    Client:
      type: object
      properties:
        client_id:
          type: string
        user_id:
          type: string
        name:
          type: string
        document:
          type: string
        register_date:
          type: string
          format: date-time
        status:
          type: boolean
    Account:
      type: object
      properties:
        account_id:
          type: string
        client_id:
          type: string
        user_id:
          type: string
        agency_id:
          type: integer
          format: uint32
        register_date:
          type: string
          format: date-time
        balance:
          type: number
          format: float64
        status:
          type: boolean
    Transfer:
      type: object
      properties:
        transfer_id:
          type: string
        account_id:
          type: string
        account_to:
          type: string
        value:
          type: number
          format: float64
        register_date:
          type: string
          format: date-time
    Deposit:
      type: object
      properties:
        deposit_id:
          type: string
        account_id:
          type: string
        client_id:
          type: string
        user_id:
          type: string
        agency_id:
          type: integer
          format: uint32
        deposit:
          type: number
          format: float64
        deposit_date:
          type: string
          format: date-time
...
    ClientReport:
      type: object
      properties:
        client_id:
          type: string
        user_id:
          type: string
        name:
          type: string
        document:
          type: string
        register_date:
          type: string
          format: date-time
        status:
          type: boolean
        accounts:
          type: array
          items:
            $ref: '#/components/schemas/Account'
        report_date:
          type: string
          format: date-time
...